+++ src/SlabSeaIceModels/slab_heat_and_tracer_fluxes.jl
@@ -1,0 +1,31 @@
+struct ConductiveFlux{K}
+    conductivity :: K
+end
+
+ConductiveFlux(FT::DataType=Float64; conductivity) = ConductiveFlux(convert(FT, conductivity))
+
+@inline function slab_internal_thermal_flux(conductive_flux::ConductiveFlux,
+                                            top_surface_temperature,
+                                            bottom_temperature,
+                                            ice_thickness)
+
+    k = conductive_flux.conductivity
+    Tu = top_surface_temperature
+    Tb = bottom_temperature
+    h = ice_thickness
+
+    return ifelse(h <= 0, zero(h), - k * (Tu - Tb) / h)
+end
+
+@inline function slab_internal_thermal_flux(i, j, grid,
+                                            top_surface_temperature::Number,
+                                            clock, fields, parameters)
+    flux = parameters.flux
+    bottom_bc = parameters.bottom_thermal_boundary_condition
+    liquidus = parameters.liquidus
+    Tu = top_surface_temperature
+    Tb = bottom_temperature(i, j, grid, bottom_bc, liquidus)
+    h = @inbounds fields.h[i, j, 1]
+    return slab_internal_thermal_flux(flux, Tu, Tb, h)
+end
+
